Changed coordinates type from int to double.

+++ src/main/java/pl/edu/agh/agents/gui/Point.java
@@ -1,19 +1,19 @@
 package pl.edu.agh.agents.gui;
 
 public class Point {
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
